captiveportal: fix ipfw rules
When we authorise a client we add it to the *_auth_(up|down) tables.
This means traffic will pass and not be forwarded, as piped traffic does
not pass through the firewall again (if net.inet.ip.fw.one_pass is set).

However, these rules are 'layer2', so when the traffic is passed it's
only passed in the layer 2 checks. Once it hits the layer 3 pfil hook
(and ipfw again) it will hit the fwd rule again, and be directed at the
captive portal webserver, despite being authorised.

Tag the authorised traffic in the layer 2 rule and allow tagged traffic
to unconditionally pass it.

Regression:	#12345